Give up on trying to handle signals with Go.
Prior patches didn't take into account that the signals can
be delivered on a different thread than where they originate.
This means it is unsafe to longjmp from the SIGINT handler
to the gets() function, because this might migrate control
from one thread to another. In Go's context, this can
in turn cause the same goroutine to resume in two threads,
causing endless confusion in the Go scheduler.

Meanwhile, also sigprocmask(2) does not configure the signal
mask globally in a multithreaded program. This means that
whatever libedit does when EL_SIGNAL is set is plainly wrong.
Experimentally, I've observed it not restoring the signal
mask properly on darwin specifically.

This patch throws the towel on trying to be too clever and
comes back to the basics:

- asks the terminal to not issue signals on ^C and ^Z.
- sets up a ^Z key handler which issues SIGTSTP, letting
  the standard Go/OS sighandler take it over, and then
  resets/redraws the prompt upon resume before returning.
  This ensures the reset properly takes place upon process
  resumption, not from a signal handler so no running
  afoul of weird darwin semantics, and on the same thread
  as libedit so that it doesn't confuse itself.
- makes the ^C key handler not do longjmp, but instead simply
  complete the line and mark the interrupt in a global
  variable to be picked up when el_gets returns. This is
  actually a better solution because it lets libedit
  clean up after itself properly.
- sets up the dumbest, no-state-needed sig handler for
  SIGCONT and SIGWINCH so that terminal interactions do
  the right thing. Experimentally, this was checked to
  not trigger weird interactions with the Go runtime.
  (As of 1.9, Go's runtime ignores those signals.)