Fix the restore of signal handlers on darwin
Prior to this patch, the signal handlers were not properly restored
upon returning from go_libedit_gets(), causing crashes if the signals
were received in Go after gets() returned and before the next call.

tl;dr: This patch fixes this problem, by ensuring the signal
configuration is properly restored.

----

On macOS/darwin, the signal trampoline (the function called by the
kernel to prepare the register context for the signal handler) is a
user-defined function *separate from the signal handler*.

Go installs its own trampoline for every signal handler, and expects
its trampoline to stay in place across cgo calls.
(On every other OS, there is no trampoline separate from the signal
handler, and Go installs its trampoline as regular signal handler.)

A first obstacle here is that the libc wrapper around the darwin
syscall (the C function with symbol `sigaction`) always forces
the trampoline to become libc's own trampoline (called `_sigtramp`),
and not Go's (called `runtime·sigtramp`).

See e.g.:
https://opensource.apple.com/source/Libc/Libc-1244.1.7/sys/sigaction.c.auto.html

So in order to configure Go's trampoline back, we cannot use
the C library wrapper, and instead we must use the low-level
syscall wrapper called `__sigaction`, which operates on a
`struct __sigaction` (with a `sa_tramp` field), instead of the
standard `struct sigaction` (which doesn't have `sa_tramp`).

(Naturally, we need to declare it ourselves.)

Another rather unfortunate obstacle, is that the `__sigaction()`
system call on macOS/darwin allows a process to *set* the
trampoline, but not *retrieve* it. This is a ... curious, badly
documented and inconvenient limitation of that syscall on darwin.

This makes our life difficult, because we want to set our own
signal handler for SIGCONT and SIGWINCH, and then "restore the
previous signal handler configuration" upon returning. So we can't
use `__sigaction`'s `oact` argument and then use it as `act` and expect
things to go well -- Go would be left with the libc's `_sigtramp`
and become confused.

So we need to proceed as follows:

- we must properly identify Go's signal trampoline function. If we
  can't identify it properly, we can't restore it, so we shouldn't
  even try changing the signal handlers.

  To populate g_sigtramp, we have no choice but to use grey
  magic. See the `unix/sigtramp` package for details.

- we must override the `sa_tramp` field in the saved `struct
  __sigaction`, not counting on the initial call to `__sigaction()` to
  set it properly for us.