Fix the broken assumption about the presence of libtinfo on linux
Prior to this patch, the LDFLAGS for linux were set to link against
libtinfo, mistakenly assuming that this would cause the resulting Go
binary to only use the subset of whatever tinfo implementation on
linux system dedicated to terminal access.

The mistake is that there is no such thing -- there is no standad
standalone libtinfo on linux, and on most linux system it is either a
symlink to libncurses, or doesn't exist at all!

This patch corrects the issue by linking against libncurss instead,
which is guaranteed to exist (it's been part of the Linux Standard
Base forever).