Remove macOS sigtramp hack
Since Go 1.11, Go no longer abuses the internal sigtramp field on macOS
when it installs its signal handlers [0]. This means that we no longer
need several layers of deviousness to restore Go's signal handlers on
Darwin.

Once this commit lands, we should add a note to the README indicating
that users who want support for Go versions early than 1.11 will need to
use the commit prior to this one.

[0]: https://github.com/golang/go/commit/ec989337c5d3203d52de1a2314813996c711fce6